requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pymysql==1.1.0
cryptography==41.0.7
django-cors-headers==4.3.0
drf-spectacular==0.26.5
PyJWT==2.8.0
python-dotenv==1.0.0
